+++ packages/php/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.0.1
+
+- Add RFBP\Producer\CollectionConsumer
+- Add RFBP\Producer\CollectionProducer
+
 # 1.0.0
 
 - Initial release : Rail, Supervisor, Client